requirements.txt
pyyaml
spotipy
